Moved FinalContext and ErroneousAndValidPartitionData classes outside SharePartitionManager and added a required synchronized block in hasNext() in ShareSessionContext
